+++ test/blockbuster/manager_test.rb
