+++ test/blockbuster/master_test.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'test_helper'
 
 describe Blockbuster::Master do
   let(:configuration) { Blockbuster::Configuration.new }
